Changed the spring to 3.2 because the <mvc:excluding> label does not support in spring 3.0.
Changed the spring to 3.2 because the <mvc:excluding> label does not
support in spring 3.0. Note that need to change all the spring model
from 3.0 to 3.2, otherwise will show some class or method can't be found
issue. And some strange issue.

When eclipse shows some method or class can't be found, you need to
check if the jar package was properly imported.